Implémentation terminée des projectiles

+++ src/TransformComponent.hpp
@@ -18,7 +18,7 @@ public:
 
 	int height = 64;
 	int width = 64;
-	int scale = 1;
+	float scale = 1.0;
 
 	float speed = 0.07;
 
@@ -26,7 +26,7 @@ public:
 
 	TransformComponent(){}
 
-	TransformComponent(float x, float y, int h, int w, int sc)
+	TransformComponent(float x, float y, int h, int w, float sc)
 	{
 		position.x = x;
 		position.y = y;
